# web_root/sample/js/swagger.yaml
swagger: '2.0'
info:
  description:
    具体实现的功能如下<br>
      1. PGP加密文件<br>
      2. 上传文件并PGP加密， 到Zurich的sftp<br>
      3. 自动识别图片文件，将图片文件转换成PDF后再PGP加密上传
  title: pgp-sftp-proxy
tags: []
host: "127.0.0.1:3333"
basePath: "/"
schemes:
- http
consumes: []
produces: []
paths:
  /encrypt:
    post:
      tags: []
      summary: PGP加密
      description: <p>对上传的文件进行PGP加密</p>
      operationId: ''
      schemes:
        - http
      consumes:
        - multipart/form-data
      produces:
        - application/pgp
      parameters:
        - name: upload
          in: formData
          description: 需要PGP加密的文件
          required: true
          type: file
          default: ''
          format: ''
        - name: key
          in: formData
          description: PGP 公匙内容
          required: true
          type: string
          default: ''
          format: ''
      responses: 
        500:
          description: API报错
          schema:
            type: object
            properties:
              status:
                type: string
                description: 状态码
              error:
                type: string
                description: 错误内容
            example:
              status: 0
              error: "未知错误"
        200:
          description: PGP文件内容
  /upload:
    post:
      tags: []
      summary: 上传文件并PGP加密
      description: <p>上传一个文件，并进行PGP加密</p>
      operationId: ''
      schemes:
        - http
      consumes:
        - multipart/form-data
      produces:
        - application/pgp
      parameters:
        - name: upload
          in: formData
          description: 需要PGP加密的文件
          required: true
          type: file
          default: ''
          format: ''
        - name: key
          in: formData
          description: PGP 公匙内容
          required: true
          type: string
          default: ''
          format: ''
        - name: deploy
          in: formData
          description: 上传文件夹
          required: true
          type: string
          enum: ["dev", "pro", "testing"]
          default: ''
          format: ''
      responses: 
        500:
          description: API报错
          example:
              status: 0
              error: "未知错误"
        200:
          description: PDF文件内容
  /multiple/upload:
    post:
      tags: []
      summary: 多文件上传及加密
      description: <p>将多个文件URL上传到sftp，并PGP加密</p>
      operationId: ''
      schemes:
        - http
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: sdadasd
          description: 需要转换成PDF的页面HTML
          required: true
          schema:
            type: object
            properties:
              files:
                type: object
                description: "待上传文件结构， filename => file url"
              key:
                type: string
                description: PGP 加密公匙
              notify: 
                type: string
                description: "通知回调，当PGP及upload 完成后，会HTTP GET 该地址，需要在response中输出 success，作为正确接收相应，不然会重试访问2次"
            example:
              files:
                "M_Article_Zurich_2.gif": "https:\/\/s3-ap-southeast-1.amazonaws.com\/s3.jetso.com\/asset\/images\/M_Article_Zurich_2.gif"
                "M_Article_Zurich_4.jpg": "https:\/\/s3-ap-southeast-1.amazonaws.com\/s3.jetso.com\/asset\/images\/M_Article_Zurich_4.jpg"
                "M_Article_Zurich_ca.gif": "https:\/\/s3-ap-southeast-1.amazonaws.com\/s3.jetso.com\/asset\/images\/M_Article_Zurich_ca.gif"
              key: "nmQENBFV2aQkBCADuEi0WB"
              notify: "http:\/\/www.baidu.com\/"
      responses: 
        500:
          description: API报错
          example:
              status: 0
              error: "未知错误"
        200:
          description: PDF文件内容
          example:
              status: 1